start of demo page work for stocks
